Move slow test related methods to lower level conftest.py
- They conflicted with astropy's test plugins which
  were imported as *.
